pyproject.toml
[tool.poetry]
name = "socket-client1"
version = "0.1.0"
description = ""
authors = ["Davee Sok <daveesok@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-socketio = {extras = ["client"], version = "^5.4.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
